summary commit

+++ controllers/contacts/remove.js
@@ -2,8 +2,10 @@ const Contact = require("../../models/contacts");
 const missingContact = require("./missingContact");
 
 const remove = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const result = await Contact.findOneAndDelete({ _id: contactId });
+  
+  const result = await Contact.findOneAndDelete({ _id: contactId, owner });
 
   !result && missingContact(contactId);
